Reduce redundant code, use common THAI2FIT_TOKENIZER from pythainlp.tokenize
