+++ docs/api/ulmfit.rst
@@ -2,25 +2,21 @@
 
 pythainlp.ulmfit
 ====================================
-The :class:`ulmfit.utils` is utils for ULMFit model.
 
 Modules
 -------
-.. autoclass:: ThaiTokenizer
 .. autofunction:: document_vector
-.. autofunction:: process_thai
 .. autofunction:: fix_html
-.. autofunction:: spec_add_spaces
-.. autofunction:: rm_useless_spaces
-.. autofunction:: rm_useless_newlines
-.. autofunction:: rm_brackets
-.. autofunction:: replace_rep_nonum
-.. autofunction:: ungroup_emoji
 .. autofunction:: lowercase_all
+.. autofunction:: merge_wgts
+.. autofunction:: process_thai
+.. autofunction:: rm_brackets
+.. autofunction:: rm_useless_newlines
+.. autofunction:: rm_useless_spaces
+.. autofunction:: remove_space
+.. autofunction:: replace_rep_after
+.. autofunction:: replace_rep_nonum
 .. autofunction:: replace_wrep_post
 .. autofunction:: replace_wrep_post_nonum
-.. autofunction:: replace_rep_after
-.. autofunction:: remove_space
-.. autofunction:: merge_wgts
-
-:members: tokenizer
+.. autofunction:: spec_add_spaces
+.. autofunction:: ungroup_emoji
